+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Additional config file locations added. See README.md for details.
+- Additional config file locations added. See README.md for details
+- Added `offset` parameter to `consume`
 - Support for basic auto completion in fish shell
 
 ## 1.1.0 - 2019-03-14