+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 1.2.0 - 2019-05-24
+
 ### Added
 - Additional config file locations added. See README.md for details
 - Added `offset` parameter to `consume`